fixxing some bugs

+++ MemoryFree.cpp
@@ -25,22 +25,22 @@ extern struct __freelist *__flp;
 
 
 /* Calculates the size of the free list */
-int freeListSize() {
+long freeListSize() {
   struct __freelist* current;
-  int total = 0;
+  long total = 0;
   for (current = __flp; current; current = current->nx) {
     total += 2; /* Add two bytes for the memory block's header  */
-    total += (int) current->sz;
+    total += (long) current->sz;
   }
   return total;
 }
 
-int freeMemory() {
-  int free_memory;
-  if ((int)__brkval == 0) {
-    free_memory = ((int)&free_memory) - ((int)&__heap_start);
+long freeMemory() {
+  long free_memory;
+  if ((long)__brkval == 0) {
+    free_memory = ((long)&free_memory) - ((long)&__heap_start);
   } else {
-    free_memory = ((int)&free_memory) - ((int)__brkval);
+    free_memory = ((long)&free_memory) - ((long)__brkval);
     free_memory += freeListSize();
   }
   return free_memory;
@@ -53,7 +53,7 @@ extern "C" {
 #include "user_interface.h"
 }
 
-int freeMemory() {
+long freeMemory() {
 	return system_get_free_heap_size();
 }
 
@@ -62,7 +62,7 @@ int freeMemory() {
 #ifdef ARDUINO_SAMD_ZERO
 extern "C" char *sbrk(int i);
  
-int freeMemory() {
+long freeMemory() {
   char stack_dummy = 0;
   return &stack_dummy - sbrk(0);
 }
